requirements.txt
python-dotenv==1.2.1
PyMySQL==1.1.2
requests==2.32.5
sqlmodel==0.0.37
